Se añade funcion de confirmacion para eliminar usuario con javascript

+++ vista/crud.php
@@ -55,7 +55,7 @@
 					 	   		<td><?php echo $value['password'] ?></td>
 					 	   		<td><?php echo getPrivilegio( $value['privilegio'] ) ?></td>
 					 	   		<td><a href="crear_usuario_form.php?id=<?php echo $value['id'] ?>" class="btn btn-success btn-sm">Editar</a></td>
-					 	   		<td><a href="eliminar_usuario.php?id=<?php echo $value['id'] ?>" class="btn btn-danger btn-sm">Eliminar</a></td>
+					 	   		<td><a href=" javascript: eliminar(confirm('¿Deseas eliminar este usuario?'),'eliminar_usuario.php?id=<?php echo $value['id'] ?>');" class="btn btn-danger btn-sm">Eliminar</a></td>
 					 	   	</tr>
 					 	   
 					 	   <?php
@@ -72,6 +72,13 @@
 
     </div>
 
+<script type="text/javascript">
+	function eliminar(confirmacion, url){
+		if (confirmacion) {
+			window.location.href = url;
+		}
+	}
+</script>
 
     <?php 
     include 'partials/footer.php';
